add custom input validator
